start writing walkthrough and refactoring to match walkthrough description
