+++ demos/react/src/TaskApp.tsx
@@ -1,12 +1,11 @@
-import React from "react";
 import { TaskComponent } from "./TaskComponent.js";
 import { Task } from "./data/schema.js";
 import { TaskFilter } from "./TaskFilter.js";
-import { TaskTable2 } from "./TaskTable2.js";
+import { TaskTable } from "./TaskTable.js";
 import { Selected, db } from "./data/DB.js";
 import { useQuery } from "@vlcn.io/materialite-react";
 
-export const TaskApp: React.FC = () => {
+export default function TaskApp() {
   const [, selectedTask] = useQuery(
     () =>
       db.appStates.stream
@@ -29,18 +28,18 @@ export const TaskApp: React.FC = () => {
     <div className="flex h-screen">
       <div className="w-3/4 bg-gray-100 overflow-y-auto">
         <TaskFilter />
-        <TaskTable2
+        <TaskTable
           onTaskClick={onTaskSelected}
-          selectedTask={(selectedTask as Selected | null)?.id || null}
+          selectedTask={selectedTask?.id || null}
         />
       </div>
       <div className="w-1/4 bg-white overflow-y-auto p-6">
         {selectedTask ? (
-          <TaskComponent taskId={(selectedTask as Selected).id} />
+          <TaskComponent taskId={selectedTask.id} />
         ) : (
           <div>Select a task to view details</div>
         )}
       </div>
     </div>
   );
-};
+}
