+++ demos/react/src/main.tsx
@@ -1,5 +1,5 @@
 import ReactDOM from "react-dom/client";
-import { TaskApp } from "./TaskApp.js";
+import TaskApp from "./TaskApp.js";
 import "./index.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(<TaskApp />);
